% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataAssetFunction.R
\name{GetDataAsset}
\alias{GetDataAsset}
\title{Gets data asset from dropbox}
\usage{
GetDataAsset(url, target_extension = ".csv")
}
\arguments{
\item{target_extension}{}
}
\description{
Gets data asset from dropbox
}
